% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_class.R
\name{predict_class}
\alias{predict_class}
\title{Predict class assignments in subgroup model.}
\usage{
predict_class(model, X, threshold = 0.5)
}
\arguments{
\item{model}{Result from calling rsmem function.}

\item{X}{Design matrix for subgroup membership model}

\item{threshold}{Observations with p > threshold are assigned to A, B otherwise.}
}
\value{
A vector of class assignments.
}
\description{
Predict class assignments in subgroup model.
}
