% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmem.R
\name{rsmem}
\alias{rsmem}
\title{Fit (regularized) subgroup mixture of experts using metaheuristics}
\usage{
rsmem(Y, Z, X, iter, swarm, algorithm, bounds, lam = NA, show_progress = T)
}
\arguments{
\item{Y}{Outcome vector.}

\item{Z}{Design matrix for outcome-level regression.}

\item{X}{Design matrix for subgroup-level regression.}

\item{iter}{Maximum number of iterations for algorithm.}

\item{swarm}{Swarm size.}

\item{algorithm}{Algorithm to use. Supports algorithms from metaheuristicOpt.}

\item{bounds}{Parameter bounds to search within. Needs to be a named list with
entries beta1_lb, beta1_ub, beta2_lb, beta2_ub, gamma_lb, gamma_ub, sigma_lb,
sigma_ub.}

\item{lam}{Value of regularization parameter.}

\item{show_progress}{Whether to show progress bar for algorithm.}
}
\value{
A named list of parameter values and objective function value.
}
\description{
Fit (regularized) subgroup mixture of experts using metaheuristics
}
\examples{
# fit to Hitters data with no regularization
Hitters = na.omit(ISLR2::Hitters)[c(-173, -241),]
y = Hitters$Salary
z = model.matrix(Salary ~ RBI, Hitters)
x = model.matrix(Salary~ ., Hitters)[,c(-1,-6)]
bounds = list(
  beta1_lb = 0,
  beta1_ub = max(Hitters$Salary),
  beta2_lb = 0,
  beta2_ub = max(Hitters$Salary),
  gamma_lb = -5,
  gamma_ub = 5,
  sigma_lb = 0.0001,
  sigma_ub = sd(Hitters$Salary)
)
set.seed(149)
res = rsmem(y,z,x, iter = 5000, swarm = 64, 'HS', bounds)
res

table(predict_class(res, x))
}
